+++ e2e/cypress/support/commands.js
@@ -47,6 +47,7 @@ const getUrl = (componentId, fixtureName = '', config = {}) => {
 const run = (app, fixtureName, plugins) => {
   cy.visit(getUrl(app, fixtureName, plugins)).then(() => {
     disableTransitions();
+    findEditorElement();
     hideAllTooltips();
   });
 };
@@ -83,6 +84,10 @@ function disableTransitions() {
 
 function hideAllTooltips() {
   cy.get('[data-id="tooltip"]', { timeout: 300000 }).invoke('hide'); //uses jquery to set display: none
+}
+
+function findEditorElement() {
+  cy.get('.DraftEditor-root', { timeout: 300000 });
 }
 
 Cypress.Commands.add('loadEditorAndViewer', (fixtureName, config) =>