+++ examples/storybook/stories/TestCase/ShakeyTwitter.js
@@ -1,17 +1,17 @@
 import React from 'react';
 import fixture from '../../../../e2e/tests/fixtures/shakey-embed.json';
-import { htmlTypeMapper } from 'wix-rich-content-plugin-html/dist/module.viewer';
-import { RichContentViewer } from 'wix-rich-content-viewer';
+import { pluginHtml } from 'wix-rich-content-plugin-html/dist/module.viewer';
+import { RicosViewer } from 'ricos-viewer';
 
 import { RichContentViewerBox, ContentState, Section, Page } from '../Components/StoryParts';
 
-const typeMappers = [htmlTypeMapper];
+const plugins = [pluginHtml()];
 export default () => {
   return (
     <Page title="Instagram Embed">
       <Section title={'Height check'}>
         <RichContentViewerBox preset="blog-preset">
-          <RichContentViewer initialState={fixture} typeMappers={typeMappers} />
+          <RicosViewer content={fixture} plugins={plugins} />
         </RichContentViewerBox>
       </Section>