+++ CHANGELOG.md
@@ -14,6 +14,10 @@
     Changes that have landed in master but are not yet released.
     Click to see more.
 
+### :bug: Bug Fix
+- `wrapper` `viewer`
+  - [#1101](https://github.com/wix-incubator/rich-content/pull/1101) fix hashtag decorator in viewer
+
 ### :house: Internal
 - `editor` `viewer`
   - [#1090](https://github.com/wix-incubator/rich-content/pull/1090) support normalize config with {disableInlineImages: false/true}
@@ -29,6 +33,8 @@
   - [#1084](https://github.com/wix-incubator/rich-content/pull/1084) external toolbar API added
 - `fullscreen`
   - [#845](https://github.com/wix-incubator/rich-content/pull/845) fullscreen image titles and design changes
+- `wrapper`
+  - [#1003](https://github.com/wix-incubator/rich-content/pull/1003) RichContentWrapper was split into RicosEditor & RicosViewer
 - `viewer`
   - [#1088](https://github.com/wix-incubator/rich-content/pull/1048) viewerAction callback in helpers is now supported, triggered on image expand (gallery & viewer)
 ### :bug: Bug Fix
@@ -39,6 +45,8 @@
 ### :house: Internal
 - `exampleApp`
   - [0f849222](https://github.com/wix-incubator/rich-content/commit/0f849222) fix translations
+- `storybook`
+  - [#1003](https://github.com/wix-incubator/rich-content/pull/1003) added story for wrapper handling of static text toolbar
   - [#1084](https://github.com/wix-incubator/rich-content/pull/1084) external plugin sidebar added
 - `storybook`
   - [#1084](https://github.com/wix-incubator/rich-content/pull/1084) initial intent example added
@@ -65,7 +73,6 @@
   - [#898](https://github.com/wix-incubator/rich-content/pull/898) text indentation
 - `viewer`
   - [#1005](https://github.com/wix-incubator/rich-content/pull/1005) add viewMode SEO to ProGallery
-
 ### :bug: Bug Fix
 - `social-modals`
   - [#1037](https://github.com/wix-incubator/rich-content/pull/1037) disable text input autocomplete